// models/apiUserModel.ts
export interface ApiUserModel {
    name: string;
    email: string;
    password: string;
    title: string;
    dayOfBirth: string;
    monthOfBirth: string; // numeric format
    yearOfBirth: string;
    firstName: string;
    lastName: string;
    company: string;
    address: string;
    address2: string;
    country: string;
    state: string;
    city: string;
    zipcode: string;
    mobileNumber: string;
}